# swagger.json
{"swagger":"2.0","info":{"description":"API de consultas aos debitos no DETRAN-ES","version":"1.0.0","title":"api-detran-debitos"},"basePath":"/","tags":[{"name":"veiculos-debitos","description":""}],"schemes":["http"],"paths":{"/veiculos/{placa}/{renavam}":{"get":{"summary":"Dados do veiculo","description":"retorna os dados do veiculo através do WebService InternetBanking","parameters":[{"name":"renavam","required":true,"in":"path","description":"Renavam do veiculo.","type":""},{"name":"placa","required":true,"in":"path","description":"Placa do veiculo","type":""}],"responses":{"200":{"description":"Retorna informações do veiculo ","schema":{"$ref":"#/definitions/VeiculoRetorno"}},"403":{"description":"Retorna uma MensagemErro"}},"tags":["veiculos-debitos"],"produces":["application/json"],"consumes":["application/json"]}},"/veiculos/{placa}/{renavam}/debitos":{"get":{"summary":"Débitos do veiculo","description":"retorna uma lista com os débitos do veiculo","parameters":[{"name":"renavam","required":true,"in":"path","description":"Renavam do veiculo","type":""},{"name":"placa","required":true,"in":"path","description":"Placa do veiculo","type":""}],"responses":{"200":{"description":"Veiculo encontrado, retorna um array de debitos","schema":{"$ref":"#/definitions/Debito"}},"403":{"description":"Retorna uma MensagemErro"}},"tags":["veiculos-debitos"],"produces":["application/json"],"consumes":["application/json"]}},"/veiculos/{placa}/{renavam}/debitos-preview":{"get":{"summary":"Prévia dos débitos do veiculo","description":"Retorna uma previa dos débitos do veiculo","parameters":[{"name":"renavam","required":true,"in":"path","description":"Renavam do veiculo","type":""},{"name":"placa","required":true,"in":"path","description":"Placa do veiculo","type":""}],"responses":{"200":{"description":"Veiculo encontrado, retorna um array de debitos","schema":{"$ref":"#/definitions/TipoDebito"}},"403":{"description":"Retorna uma MensagemErro"}},"tags":["veiculos-debitos"],"produces":["application/json"],"consumes":["application/json"]}},"/veiculos/{placa}/{renavam}/debitos-tipo/{tipo_debito}":{"get":{"summary":"Lista de débitos filtrado por tipo","description":"Retorna uma lista de débitos do veiculo filtrada pelo tipo:                  LICENCIAMENTOATUAL, LICENCIAMETO ANTERIOR, IPVA, IPVAANTERIOR, DPVT, DPVATANTERIOR, MULTA","parameters":[{"name":"tipo_debito","required":true,"in":"path","description":"Tipo de debitos","type":""},{"name":"renavam","required":true,"in":"path","description":"Renavam do veiculo","type":""},{"name":"placa","required":true,"in":"path","description":"Placa do veiculo","type":""}],"responses":{"200":{"description":"Veiculo encontrado, retorna um array de debitos","schema":{"$ref":"#/definitions/Debito"}},"403":{"description":"Retorna uma MensagemErro"}},"tags":["veiculos-debitos"],"produces":["application/json"],"consumes":["application/json"]}},"/veiculos/{placa}/{renavam}/debitos/guia":{"get":{"summary":"Gerar GRU de todos os débitos","description":"Retornar uma GRU com todos os debitos ","parameters":[{"name":"renavam","required":true,"in":"path","description":"Renavam do veiculo","type":""},{"name":"placa","required":true,"in":"path","description":"Placa do veiculo","type":""}],"responses":{"200":{"description":"Veiculo encontrado, retorna o um array de itens e o pdf do boleto, em base64","schema":{"$ref":"#/definitions/GerarGuiaRetorno"}},"403":{"description":"Retorna uma MensagemErro"}},"tags":["veiculos-debitos"],"produces":["application/json"],"consumes":["application/json"]}},"/veiculos/{placa}/{renavam}/debitos/guia/{tipo_debito}/{listaIDs}":{"get":{"summary":"Gerar GRU de alguns debitos","description":"Retornar uma guia para pagamento dos debitos requisitados.","parameters":[{"name":"listaIDs","required":true,"in":"path","description":"Lista de IDs de debitos, separados por virgula","type":""},{"name":"tipo_debito","required":true,"in":"path","description":"tipo do debito","type":""},{"name":"renavam","required":true,"in":"path","description":"Renavam do veiculo","type":""},{"name":"placa","required":true,"in":"path","description":"Placa do veiculo","type":""}],"responses":{"200":{"description":"Veiculo encontrado, retorna o um array de itens e o pdf em base64 do boleto","schema":{"$ref":"#/definitions/DebitoRetorno"}},"403":{"description":"Retorna uma MensagemErro"}},"tags":["veiculos-debitos"],"produces":["application/json"],"consumes":["application/json"]}},"/veiculos/debitos/getGuia/{codigoBarra}":{"get":{"summary":"Guia PDF","description":"Retornar um pdf da guia para pagamento.","parameters":[{"name":"codigoBarra","required":true,"in":"path","description":"Código de barras da guia gerada","type":""}],"responses":{"200":{"description":"Guia encontrada, retorna uma guia em pdf ","schema":{"$ref":"#/definitions/DebitoRetorno"}},"403":{"description":"Retorna uma MensagemErro"}},"tags":["veiculos-debitos"],"produces":["application/json"],"consumes":["application/json"]}}},"definitions":{"VeiculoRetorno":{"type":"object","properties":{"placa":{"type":"string"},"modelo":{"type":"string"},"renavam":{"type":"number"},"mensagemErro":{"type":"string"}},"required":["placa","modelo","renavam","mensagemErro"]},"Debito":{"type":"object","properties":{"classe":{"type":"string","enum":["Licenciamento","Registro Veículo","IPVA","Seguro DPVAT","Multas","Vistoria","Certidões","Apreensão","Credenciamento","Diversos Veículos","Parcelamento IPVA","Placas","Pátio","Registro de Contrato"]},"codigoServico":{"type":"number"},"descricaoServico":{"type":"string"},"dataVencimento":{"type":"string"},"exercicio":{"type":"number"},"idDebito":{"type":"number"},"parcela":{"type":"number"},"placa":{"type":"string"},"valorAtualizadoFranquia":{"type":"number"},"flagDpvatAnterior":{"type":"number"},"dpvatCotas":{"type":"string"},"flagDpvatExercicio":{"type":"number"},"flagIpvaAnterior":{"type":"number"},"flagIpvaExercicio":{"type":"number"},"flagIpvaParcelamento":{"type":"number"},"flagLicenciamentoAnterior":{"type":"number"},"flagLicenciamentoExercicio":{"type":"number"},"flagMultas":{"type":"number"},"flagTaxaEspecial":{"type":"number"},"flagTaxaPatio":{"type":"number"},"flagTaxaServico":{"type":"number"},"ipvaCotas":{"type":"string"}},"required":["classe","codigoServico","descricaoServico","dataVencimento","exercicio","idDebito","parcela","placa","valorAtualizadoFranquia","flagDpvatAnterior","dpvatCotas","flagDpvatExercicio","flagIpvaAnterior","flagIpvaExercicio","flagIpvaParcelamento","flagLicenciamentoAnterior","flagLicenciamentoExercicio","flagMultas","flagTaxaEspecial","flagTaxaPatio","flagTaxaServico","ipvaCotas"]},"TipoDebito":{"type":"object","properties":{"temLicenciamentoAtual":{"type":"string"},"temLicenciamentoAnterior":{"type":"string"},"temDPVAT":{"type":"string"},"temIPVA":{"type":"string"},"temMulta":{"type":"string"},"temIPVAAnterior":{"type":"string"},"temDPVATAnterior":{"type":"string"},"mensagemErro":{"type":"string"}},"required":["temLicenciamentoAtual","temLicenciamentoAnterior","temDPVAT","temIPVA","temMulta","temIPVAAnterior","temDPVATAnterior","mensagemErro"]},"ItemGuia":{"type":"object","properties":{"linhaDigitavel":{"type":"string"},"codigoBarra":{"type":"string"},"valorGuia":{"type":"number"},"postagem":{"type":"boolean"},"nome":{"type":"string"},"placa":{"type":"string"},"renavam":{"type":"number"},"marca":{"type":"string"},"nossoNumero":{"type":"string"},"vencimentoGuia":{"type":"string"},"tipoRegistro":{"type":"number"},"classe":{"type":"string","enum":["Licenciamento","Registro Veículo","IPVA","Seguro DPVAT","Multas","Vistoria","Certidões","Apreensão","Credenciamento","Diversos Veículos","Parcelamento IPVA","Placas","Pátio","Registro de Contrato"]},"descricaoServico":{"type":"string"},"dataVencimento":{"type":"string"},"valorVencimento":{"type":"number"},"valorCorrigido":{"type":"number"},"valorDesconto":{"type":"number"},"valorJuros":{"type":"number"},"valorMulta":{"type":"number"},"valorAtualizadoFranquia":{"type":"number"},"dataAutuacao":{"type":"string"},"horaAutuacao":{"type":"string"},"valorAuto":{"type":"number"}},"required":["linhaDigitavel","codigoBarra","valorGuia","postagem","nome","placa","renavam","marca","nossoNumero","vencimentoGuia","tipoRegistro","classe","descricaoServico","dataVencimento","valorVencimento","valorCorrigido","valorDesconto","valorJuros","valorMulta","valorAtualizadoFranquia","dataAutuacao","horaAutuacao","valorAuto"]},"GerarGuiaRetorno":{"type":"object","properties":{"itensGuia":{"type":"array","items":{"$ref":"#/definitions/ItemGuia"}},"mensagemErro":{"type":"string"},"guiaPDF":{"type":"string"}},"required":["itensGuia","mensagemErro","guiaPDF"]},"DebitoRetorno":{"type":"object","properties":{"debitos":{"type":"array","items":{"$ref":"#/definitions/Debito"}},"mensagemErro":{"type":"string"}},"required":["debitos","mensagemErro"]}}}